Moving password strength evaluation code to separete file
