+++ index.html
@@ -93,6 +93,7 @@
     </div>
     <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/zxcvbn@4.4.2/dist/zxcvbn.min.js"></script>
+    <script src="src/levelindicators.js"></script>
     <script src="src/script.js" type="module"></script>
   </body>
 </html>
